requirements/myfile.txt
sentry-sdk 
django-anymail